move two fixtures back to bgp aggregate helprs file
